requirements.txt
streamlit
google-genai
supabase
python-dotenv
Pillow
gotrue